Improve mouse event handling.
Check if a left-click is part of a mouse drag, or just a (ctrl-)click.  Only
pass on the mouse event if it's not part of a drag movement.

Check if drag is part of moving a marker, or of it's a selection drag.

Remove unnecessary scrollContentsBy() override in WaveView.